requirements.txt
cycler==0.10.0
dlib==19.22.0
imutils==0.5.4
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
